+++ build.zig
@@ -28,7 +28,7 @@ pub fn build(b: *std.Build) void {
     // to our consumers. We must give it a name because a Zig package can expose
     // multiple modules and consumers will need to be able to specify which
     // module they want to access.
-    const mod = b.addModule("TheHashTable", .{
+    const mod = b.addModule("verztable", .{
         // The root source file is the "entry point" of this module. Users of
         // this module will only be able to access public declarations contained
         // in this file, which means that if you have declarations that you
@@ -73,12 +73,12 @@ pub fn build(b: *std.Build) void {
             // List of modules available for import in source files part of the
             // root module.
             .imports = &.{
-                // Here "TheZigHashTable" is the name you will use in your source code to
-                // import this module (e.g. `@import("TheZigHashTable")`). The name is
+                // Here "verztable" is the name you will use in your source code to
+                // import this module (e.g. `@import("verztable")`). The name is
                 // repeated because you are allowed to rename your imports, which
                 // can be extremely useful in case of collisions (which can happen
                 // importing modules from different packages).
-                .{ .name = "TheHashTable", .module = mod },
+                .{ .name = "verztable", .module = mod },
             },
         }),
     });